requirements.txt
amazon.ion==0.9.1
amqp==5.1.1
asgiref==3.4.1
astroid==2.11.3
async-timeout==4.0.2
autopep8==1.6.0
billiard==3.6.4.0
black==22.3.0
boto3==1.22.5
botocore==1.25.5
celery==5.2.6
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
Deprecated==1.2.13
dill==0.3.4
Django==3.2.13
django-autocomplete-light==3.9.4
django-axes==5.32.0
django-background-tasks==1.2.5
django-compat==1.0.15
django-crispy-forms==1.14.0
django-elasticache==1.0.3
django-elastipymemcache==2.0.4
django-ipware==4.0.2
django-localflavor==3.1
django-nose==1.4.7
django-pymemcache==1.0.0
django-redis-sessions==0.6.2
django-rest-swagger==2.2.0
django-searchable-select==1.5.0
django-selectable==1.3.0
django-session-security==2.6.6
django-silk==4.2.0
django-storages==1.12.3
django-timezone-utils==0.13
django-user-agents==0.4.0
djangorestframework==3.13.1
docutils==0.18
dparse==0.5.1
drf-yasg==1.20.0
flake8==4.0.1
gprof2dot==2021.2.21
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
ionhash==1.2.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
jmespath==1.0.0
kombu==5.2.4
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
model-bakery==1.4.0
mypy-extensions==0.4.3
nose==1.3.7
openapi-codec==1.3.2
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.2
pycodestyle==2.8.0
pyflakes==2.4.0
Pygments==2.12.0
pylibmc==1.6.1
pylint==2.13.7
pymemcache==3.5.1
pyparsing==3.0.8
pyqldb==3.2.2
python-dateutil==2.8.2
python-memcached==1.59
python-stdnum==1.17
pytz==2021.3
PyYAML==6.0
redis==4.2.2
regex==2022.4.24
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.5.2
safety==1.10.3
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
typed-ast==1.5.1
typing-extensions==4.0.1
ua-parser==0.10.0
uritemplate==4.1.1
urllib3==1.26.7
user-agents==2.2.0
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.13.3
